Move isAmbiguousThemeFilterTerm
